+++ include/mruby_integration/models/rectangle.hpp
@@ -1,8 +1,19 @@
 #pragma once
 
 #include "mruby.h"
+#include "raylib.h"
 
 extern RClass* Rectangle_class;
+
+struct Rektangle
+{
+  Rectangle* rectangle;
+  Color* colour;
+  Color* outline;
+  float thickness;
+  float roundness;
+  int segments;
+};
 
 void
 append_models_Rectangle(mrb_state*);