+++ test/models/camera2d_test.rb
@@ -79,11 +79,11 @@ class Test
       end
 
       def test_draw
-        rectangle = Rectangle.new(2, 2, 6, 6)
+        rectangle = Rectangle[2, 2, 6, 6, Colour::RED]
         camera = Camera2D.new
 
         camera.draw do
-          rectangle.draw(colour: Colour::RED)
+          rectangle.draw
         end
 
         assert_called [
@@ -98,12 +98,12 @@ class Test
         camera.offset.y = -3
 
         camera.draw do
-          draw_rectangle_rec(rectangle, Colour::RED)
+          rectangle.draw
         end
 
         assert_called [
           "(BeginMode2D) { camera: { offset.x: -2.000000 offset.y: -3.000000 target.x: 0.000000 target.y: 0.000000 rotation: 0.000000 zoom: 1.000000 } }",
-          "(DrawRectangleRec) { rec: { x: 2.000000 y: 2.000000 width: 6.000000 height: 6.000000 } color: { r: 230 g: 41 b: 55 a: 255 } }",
+          "(DrawRectanglePro) { rec: { x: 2.000000 y: 2.000000 width: 6.000000 height: 6.000000 } origin: { x: 0.000000 y: 0.000000 } rotation: 0.000000 color: { r: 230 g: 41 b: 55 a: 255 } }",
           "(EndMode2D) { }"
         ]
       end